Fixed IO futures/stream not hold a reference
Since we have an Arc<Drive> anyway, instead of using a reference to a
Reader/Writer for async io futures, we instead pass the drive as an arc.
This simplifies the useability of these futures/streams.

This has unfortunate that the Flush future now has a `D:Decode` generic
requirment on its type.  Because the kernel callback is shared between
Read and Write overlapped IO, the shared state in the futures/streams
has both Read and Write state in there. (IE the Flush Future (used when
writing) has shared state with the Decoder, used when reading).

This coupling seems required as the callback itself from the kernel will
couple the read and write.  We can sever this coupling but need to use
unique Arc pointers. so in favor or using less heap we share the same
arc instead of 3 arcs

Signed-off-by: Thomas Chiantia <thomas.chiantia@gmail.com>