refactor device
We're dropping the msft_runtime and just focusing on service stuff. We
migrate a the wait module from the runtime crate into here. We also
refactor the FilterForIds into a smarter Tracking StreamExt which emits
PlugEvents with futures that resolve when Unplugged

The DeviceListener is now a handle instead of a stream and the handle
can be used to trigger additional plug events

Signed-off-by: Thomas Chiantia <thomas.chiantia@gmail.com>